first step at unifying bip-wallet by breaking out common code between browser and desktop
