+++ desktop/src/Desktop/Setup.hs
@@ -9,7 +9,7 @@
 {-# LANGUAGE TypeFamilies #-}
 
 -- | Wallet setup screens
-module Desktop.Setup (WalletExists (..), runSetup, splashLogo, setupDiv, setupClass, checkPassword) where
+module Desktop.Setup (runSetup, checkPassword) where
 
 import Control.Lens ((<>~), (??), (^.), _1, _2, _3)
 import Control.Error (hush)
@@ -48,51 +48,9 @@ import Frontend.UI.Dialogs.ChangePassword (minPasswordLength)
 import Frontend.UI.Widgets.Helpers (imgWithAlt)
 import Frontend.UI.Widgets
 import Frontend.Setup.Widgets
+import Frontend.Setup.Common
 import Frontend.Crypto.Password
 import Obelisk.Generated.Static
-
--- | Used for changing the settings in the passphrase widget.
-data PassphraseStage
-  = Setup
-  | Recover
-  deriving (Show, Eq)
-
--- | Wrapper for the index of the word in the passphrase
-newtype WordKey = WordKey { _unWordKey :: Int }
-  deriving (Show, Eq, Ord, Enum)
-
--- | Setup stage
-data WalletScreen
-  = WalletScreen_Password
-  | WalletScreen_PrePassphrase
-  | WalletScreen_CreatePassphrase
-  | WalletScreen_VerifyPassphrase
-  | WalletScreen_RecoverPassphrase
-  | WalletScreen_RecoverImport
-  | WalletScreen_SplashScreen
-  | WalletScreen_Done
-  deriving (Show, Eq, Ord)
-
-setupScreenClass :: WalletScreen -> Text
-setupScreenClass = setupClass . T.toLower . tshow
-
-wordsToPhraseMap :: [Text] -> Map.Map WordKey Text
-wordsToPhraseMap = Map.fromList . zip [WordKey 1 ..]
-
--- setupClass :: Text -> Text
--- setupClass = mappend "setup__"
-
--- setupDiv :: DomBuilder t m => Text -> m a -> m a
--- setupDiv t = divClass (setupClass t)
-
--- setupCheckbox
---   :: (DomBuilder t m, PostBuild t m)
---   => Bool
---   -> CheckboxConfig t
---   -> m ()
---   -> m (Checkbox t)
--- setupCheckbox initialValue cfg inner = elClass "div" (setupClass "checkbox")
---   $ uiCheckbox def initialValue cfg inner
 
 mkPhraseMapFromMnemonic
   :: forall mw.
@@ -111,88 +69,6 @@ baToText = T.decodeUtf8 . BA.pack . BA.unpack
 
 textTo :: IsString a => Text -> a
 textTo = fromString . T.unpack
-
-tshow :: Show a => a -> Text
-tshow = T.pack . show
-
--- setupForm :: forall t m a. (DomBuilder t m, PostBuild t m) => Text -> Text -> Dynamic t Bool -> m a -> m (Event t (), a)
--- setupForm cls lbl disabled = uiForm cfg $ setupDiv cls $ void $ confirmButton (def & uiButtonCfg_disabled .~ disabled) lbl
---   where cfg = def & elementConfig_initialAttributes .~ ("class" =: setupClass "form")
-
--- restoreForm :: (DomBuilder t m, PostBuild t m) => Dynamic t Bool -> m a -> m (Event t (), a)
--- restoreForm = setupForm "recover-restore-button" "Restore"
-
--- continueForm :: (DomBuilder t m, PostBuild t m) => Dynamic t Bool -> m a -> m (Event t (), a)
--- continueForm = setupForm "continue-button" "Continue"
-
--- | Wallet logo
--- kadenaWalletLogo :: DomBuilder t m => m ()
--- kadenaWalletLogo = divClass "logo" $ do
---   elAttr "img" ("src" =: static @"img/kadena_blue_logo.png" <> "class" =: setupClass "kadena-logo") blank
---   elClass "div" "chainweaver" $ text "Chainweaver"
---   elClass "div" "by-kadena" $ text "by Kadena"
-
-type SetupWF t m = Workflow t m
-  (WalletScreen
-  , Event t (Crypto.XPrv, Password, Bool)
-  , Event t () -- ExitSetup Event
-  )
-
-finishSetupWF :: (Reflex t, Applicative m) => WalletScreen -> a -> m ((WalletScreen, Event t x, Event t ()), a)
-finishSetupWF ws = pure . (,) (ws, never, never)
-
--- Make this take a list of the current progress instead so we can maintain
--- the list of how much we have done so far.
---
-walletSetupRecoverHeader
-  :: DomBuilder t m
-  => WalletScreen
-  -> m ()
-walletSetupRecoverHeader currentScreen = setupDiv "workflow-header" $ do
-  unless (currentScreen `elem` [WalletScreen_RecoverPassphrase, WalletScreen_RecoverImport, WalletScreen_SplashScreen]) $ do
-    elClass "ol" (setupClass "workflow-icons") $ do
-      faEl "1" "Password" WalletScreen_Password
-      line WalletScreen_PrePassphrase
-      faEl "2" "Recovery" WalletScreen_CreatePassphrase
-      line WalletScreen_VerifyPassphrase
-      faEl "3" "Verify" WalletScreen_VerifyPassphrase
-
-  where
-    addActive sid =
-      if isActive sid then
-        ["active"]
-      else
-        []
-
-    line sid =
-      elClass "div" (T.unwords $ setupClass "workflow-icon": setupClass "header-line" : addActive sid) blank
-
-    isActive sid = sid `elem` (progress currentScreen)
-    progress WalletScreen_Password =
-      []
-    progress WalletScreen_PrePassphrase =
-      [WalletScreen_Password]
-    progress WalletScreen_CreatePassphrase =
-      [WalletScreen_Password, WalletScreen_PrePassphrase]
-    progress WalletScreen_VerifyPassphrase =
-      [WalletScreen_Password, WalletScreen_PrePassphrase, WalletScreen_CreatePassphrase]
-    progress WalletScreen_Done =
-      [WalletScreen_Password, WalletScreen_PrePassphrase, WalletScreen_CreatePassphrase, WalletScreen_VerifyPassphrase]
-    progress _ = []
-
-    faEl n lbl sid =
-      elClass "li" (setupClass "workflow-icon" <> if isActive sid then " active" else T.empty) $ do
-        elClass "div" (setupClass "workflow-icon-circle" <> " " <> setupClass ("workflow-screen-" <> T.toLower lbl)) $
-          setupDiv "workflow-icon-symbol" $ do
-            if isActive sid
-              then elClass "i" ("fa fa-check fa-lg fa-inverse " <> setupClass "workflow-icon-active") blank
-              else text n
-            setupDiv "workflow-icon-label" $ text lbl
-
-data WalletExists
-  = WalletExists_Yes
-  | WalletExists_No
-  deriving (Show, Eq)
 
 runSetup
   :: forall t m
@@ -236,13 +112,6 @@ runSetup fileFFI showBackOverride walletExists = setupDiv "fullscreen" $ mdo
       else
         setupScreenClass ws
 
-splashLogo :: DomBuilder t m => m ()
-splashLogo = do
-  elAttr "div"
-    (  "style" =: ("background-image: url(" <> (static @"img/Wallet_Graphic_1.png") <> ");")
-    <> "class" =: setupClass "splash-bg"
-    ) kadenaWalletLogo
-
 splashScreen
   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m
      , PostBuild t m, MonadJSM (Performable m), TriggerEvent t m, HasStorage (Performable m)
@@ -252,7 +121,7 @@ splashScreen
   => WalletExists
   -> FileFFI t m
   -> Event t ()
-  -> SetupWF t m
+  -> SetupWF Crypto.XPrv t m
 splashScreen walletExists fileFFI eBack = selfWF
   where
     selfWF = Workflow $ setupDiv "splash" $ do
@@ -291,15 +160,12 @@ data BIP39PhraseError
   | BIP39PhraseError_InvalidPhrase
   | BIP39PhraseError_PhraseIncomplete
 
-passphraseLen :: Int
-passphraseLen = 12
-
 sentenceToSeed :: Crypto.ValidMnemonicSentence mw => Crypto.MnemonicSentence mw -> Crypto.Seed
 sentenceToSeed s = Crypto.sentenceToSeed s Crypto.english ""
 
 restoreBipWallet
   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m), TriggerEvent t m)
-  => SetupWF t m -> Event t () -> SetupWF t m
+  => SetupWF Crypto.XPrv t m -> Event t () -> SetupWF Crypto.XPrv t m
 restoreBipWallet backWF eBack = Workflow $ do
   el "h1" $ text "Recover your wallet"
 
@@ -363,9 +229,9 @@ restoreFromImport
      )
   => WalletExists
   -> FileFFI t m
-  -> SetupWF t m
+  -> SetupWF Crypto.XPrv t m
   -> Event t ()
-  -> SetupWF t m
+  -> SetupWF Crypto.XPrv t m
 restoreFromImport walletExists fileFFI backWF eBack = nagScreen
   where
     nagMsgs = case walletExists of
@@ -454,70 +320,9 @@ restoreFromImport walletExists fileFFI backWF eBack = nagScreen
     nonEmptyPassword "" = Nothing
     nonEmptyPassword pw = Just (Password pw)
 
-passphraseWordElement
-  :: (DomBuilder t m, PostBuild t m, MonadFix m)
-  => Dynamic t PassphraseStage
-  -> Bool
-  -> WordKey
-  -> Dynamic t Text
-  -> m (Event t Text)
-passphraseWordElement currentStage exposeEmpty k wrd = do
-  rec
-    let eFocused = leftmost
-          [ eInputFocused
-          , True <$ domEvent Mouseenter elt
-          , True <$ domEvent Mousemove elt
-          , False <$ domEvent Mouseleave elt
-          ]
-
-    (elt, (val, eInputFocused, eInput)) <- elClass' "div" (setupClass "passphrase-widget-elem-wrapper") $ do
-      pb <- getPostBuild
-
-      setupDiv "passphrase-widget-key-wrapper" $
-        text (showWordKey k)
-
-      let
-        eExposed = attachWith (\v f -> T.null v || f) (current val) eFocused
-
-        commonClass cls exposed = "input " <> (setupClass cls <> bool "" (" " <> setupClass cls <> "--exposed") exposed)
-        commonAttrs cls exposed =
-          "type" =: "text" <>
-          "size" =: "8" <>
-          "class" =: commonClass cls exposed
-
-      setupDiv "passphrase-widget-word-wrapper" $ do
-        void . uiInputElement $ def
-          & inputElementConfig_initialValue .~ "********"
-          & initialAttributes .~ (commonAttrs "passphrase-widget-word-hider" exposeEmpty <> "disabled" =: "true" <> "tabindex" =: "-1")
-          & modifyAttributes <>~ (("class" =:) . Just . commonClass "passphrase-widget-word-hider" <$> eExposed)
-
-        inputElt <- uiInputElement $ def
-          & inputElementConfig_setValue .~ (current wrd <@ pb)
-          & initialAttributes .~ commonAttrs "passphrase-widget-word" exposeEmpty
-          & modifyAttributes <>~ fold
-            [ (("readonly" =:) . canEditOnRecover <$> current currentStage <@ pb)
-            , (("class" =:) . Just . commonClass "passphrase-widget-word" <$> eExposed)
-            ]
-
-        pure (_inputElement_value inputElt, updated $ _inputElement_hasFocus inputElt, _inputElement_input inputElt)
-  pure (T.strip <$> eInput)
-  where
-    canEditOnRecover Recover = Nothing
-    canEditOnRecover Setup = Just "true"
-
-passphraseWidget
-  :: (DomBuilder t m, MonadFix m, MonadHold t m, PostBuild t m)
-  => Dynamic t (Map.Map WordKey Text)
-  -> Dynamic t PassphraseStage
-  -> Bool
-  -> m (Event t (Map.Map WordKey Text))
-passphraseWidget dWords dStage exposeEmpty = do
-  setupDiv "passphrase-widget-wrapper" $
-    listViewWithKey dWords (passphraseWordElement dStage exposeEmpty)
-
 createNewWallet
   :: forall t m. (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m), TriggerEvent t m)
-  => SetupWF t m -> Event t () -> SetupWF t m
+  => SetupWF Crypto.XPrv t m -> Event t () -> SetupWF Crypto.XPrv t m
 createNewWallet backWF eBack = selfWF
   where
     selfWF = Workflow $  do
@@ -536,7 +341,7 @@ createNewWallet backWF eBack = selfWF
           dynText =<< holdDyn "Generating your mnemonic..." eGenError
           pure never
 
-        proceed :: Crypto.MnemonicSentence 12 -> m (Event t (SetupWF t m))
+        proceed :: Crypto.MnemonicSentence 12 -> m (Event t (SetupWF Crypto.XPrv t m))
         proceed mnem = mdo
           (ePwSubmit, dPassword) <- continueForm (fmap isNothing dPassword) $ do
             holdDyn Nothing =<< setPassword (pure $ sentenceToSeed mnem)
@@ -551,30 +356,13 @@ createNewWallet backWF eBack = selfWF
         , switchDyn dContinue
         ]
 
-walletSplashWithIcon :: DomBuilder t m => m () -> m ()
-walletSplashWithIcon w = do
-  elAttr "div" (
-    "style" =: ("background-image: url(" <> (static @"img/Wallet_Graphic_1.png") <> ");")
-    <> "class" =: (setupClass "splash-bg " <> setupClass "done-splash-bg")
-    ) $ do
-      w
-      elAttr "img" (
-        "src" =: static @"img/Wallet_Icon_Highlighted_Blue.svg" <>
-        "class" =: setupClass "wallet-blue-icon"
-        ) blank
-
-stackFaIcon :: DomBuilder t m => Text -> m ()
-stackFaIcon icon = elClass "span" "fa-stack fa-5x" $ do
-  elClass "i" "fa fa-circle fa-stack-2x" blank
-  elClass "i" ("fa " <> icon <> " fa-stack-1x fa-inverse") blank
-
 precreatePassphraseWarning
   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m))
-  => SetupWF t m
+  => SetupWF Crypto.XPrv t m
   -> Event t ()
   -> (Crypto.XPrv, Password)
   -> Crypto.MnemonicSentence 12
-  -> SetupWF t m
+  -> SetupWF Crypto.XPrv t m
 precreatePassphraseWarning backWF eBack (rootKey, password) mnemonicSentence = Workflow $ mdo
   (eContinue, dUnderstand) <- continueForm (fmap not dUnderstand) $ do
 
@@ -602,30 +390,14 @@ precreatePassphraseWarning backWF eBack (rootKey, password) mnemonicSentence = W
   where
     line = el "div" . text
 
-doneScreen
-  :: (DomBuilder t m, PostBuild t m)
-  => (Crypto.XPrv, Password)
-  -> SetupWF t m
-doneScreen (rootKey, passwd) = Workflow $ do
-  walletSplashWithIcon blank
-
-  el "h1" $ text "Wallet Created"
-
-  eContinue <- setupDiv "continue-button" $
-    confirmButton def "Done"
-
-  pure ( (WalletScreen_Done, (rootKey, passwd, True) <$ eContinue, never)
-       , never
-       )
-
 -- | UI for generating and displaying a new mnemonic sentence.
 createNewPassphrase
   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m))
-  => SetupWF t m
+  => SetupWF Crypto.XPrv t m
   -> Event t ()
   -> (Crypto.XPrv, Password)
   -> Crypto.MnemonicSentence 12
-  -> SetupWF t m
+  -> SetupWF Crypto.XPrv t m
 createNewPassphrase backWF eBack (rootKey, password) mnemonicSentence = selfWF
   where
     selfWF = Workflow $ mdo
@@ -657,12 +429,12 @@ createNewPassphrase backWF eBack (rootKey, password) mnemonicSentence = selfWF
 -- choose the words in the right order.
 confirmPhrase
   :: (DomBuilder t m, MonadFix m, MonadHold t m, PostBuild t m)
-  => SetupWF t m
+  => SetupWF Crypto.XPrv t m
   -> Event t ()
   -> (Crypto.XPrv, Password)
   -> Crypto.MnemonicSentence 12
   -- ^ Mnemonic sentence to confirm
-  -> SetupWF t m
+  -> SetupWF Crypto.XPrv t m
 confirmPhrase backWF eBack (rootKey, password) mnemonicSentence = Workflow $ mdo
   (continue, done) <- continueForm (fmap not done) $ do
     el "h1" $ text "Verify Recovery Phrase"
